+++ protected/humhub/modules/directory/views/directory/groups.php
@@ -1,8 +1,9 @@
 <?php
+/* @var $this \yii\web\View */
+/* @var $groups humhub\modules\user\models\Group[] */
 
-use yii\helpers\Url;
 use yii\helpers\Html;
-
+use humhub\modules\directory\widgets\GroupUsers;
 ?>
 <div class="panel panel-default groups">
 
@@ -12,23 +13,9 @@ use yii\helpers\Html;
 
     <div class="panel-body">
         <?php foreach ($groups as $group) : ?>
-            <?php $userCount = $group->getGroupUsers()->count(); ?>
-
-            <?php if ($userCount != 0) : ?>
-                <h1><?php echo Html::encode($group->name); ?></h1>
-                <?php foreach ($group->getUsers()->active()->limit(30)->joinWith('profile')->orderBy(['profile.lastname' => SORT_ASC])->all() as $user) : ?>
-                    <a id="<?php echo $user->guid; ?>" href="<?php echo $user->getUrl(); ?>">
-                        <img data-toggle="tooltip" data-placement="top" title=""
-                             data-original-title="<?php echo Html::encode($user->displayName); ?>"
-                             src="<?php echo $user->getProfileImage()->getUrl(); ?>" class="img-rounded tt img_margin"
-                             height="40"
-                             width="40" alt="40x40" data-src="holder.js/40x40" style="width: 40px; height: 40px;"/></a>
-                    <?php endforeach; ?>
-                    <?php if ($userCount >= 30) : ?>
-                        <?php echo Html::a(Yii::t('DirectoryModule.base', "show all members"), Url::to(['/directory/directory/members', 'keyword' => '', 'groupId' => $group->id])); ?>
-                    <?php endif; ?>
-                <hr>
-            <?php endif; ?>
+            <h1><?= Html::encode($group->name); ?></h1>
+            <?= GroupUsers::widget(['group' => $group]); ?>
+            <hr />
         <?php endforeach; ?>
     </div>